% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{get_enrolment}
\alias{get_enrolment}
\title{Get enrolment by division}
\usage{
get_enrolment(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of enrolment figures by division.
}
\description{
Get enrolment by division
}
\examples{
\dontrun{
get_enrolment(2022)
}
}
