% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{get_fp}
\alias{get_fp}
\title{Get first preference votes by division}
\usage{
get_fp(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of first preference votes by division.
}
\description{
Get first preference votes by division
}
\examples{
\dontrun{
get_fp(2022)
}
}
