% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{get_tpp_by_booth}
\alias{get_tpp_by_booth}
\title{Get two-party preferred votes by polling place}
\usage{
get_tpp_by_booth(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of TPP votes by polling place.
}
\description{
Get two-party preferred votes by polling place
}
\examples{
\dontrun{
get_tpp_by_booth(2022)
}
}
