% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{get_turnout}
\alias{get_turnout}
\title{Get turnout by division}
\usage{
get_turnout(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of turnout figures by division.
}
\description{
Get turnout by division
}
\examples{
\dontrun{
get_turnout(2022)
}
}
