% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{list_elections}
\alias{list_elections}
\title{List all available federal elections}
\usage{
list_elections()
}
\value{
A data frame with election year, date, event ID, and type.
}
\description{
List all available federal elections
}
\examples{
list_elections()
}
