% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referendum.R
\name{list_referendums}
\alias{list_referendums}
\title{List available referendums}
\usage{
list_referendums()
}
\value{
A data frame with one row per referendum with results published
on the AEC tally room, including columns \code{year}, \code{date}, \code{event_id},
and \code{description}.
}
\description{
List available referendums
}
\examples{
list_referendums()
}
