% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{get_candidates}
\alias{get_candidates}
\title{Get all candidates for an election}
\usage{
get_candidates(year, chamber = "house", refresh = FALSE)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}

\item{chamber}{"house" or "senate".}

\item{refresh}{If \code{TRUE}, re-download from the AEC even if a cached copy
exists.}
}
\value{
A tidy data frame of candidates.
}
\description{
Get all candidates for an election
}
\examples{
\donttest{
op <- options(readaec.cache_dir = tempdir())
get_candidates(2022)
get_candidates(2022, chamber = "senate")
options(op)
}
}
