% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{get_candidates}
\alias{get_candidates}
\title{Get all candidates for an election}
\usage{
get_candidates(year, chamber = "house")
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}

\item{chamber}{"house" or "senate".}
}
\value{
A tidy data frame of candidates.
}
\description{
Get all candidates for an election
}
\examples{
\donttest{
get_candidates(2022)
get_candidates(2022, chamber = "senate")
}
}
