% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{get_members_elected}
\alias{get_members_elected}
\title{Get members elected to the House of Representatives}
\usage{
get_members_elected(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of elected members.
}
\description{
Get members elected to the House of Representatives
}
\examples{
\donttest{
op <- options(readaec.cache_dir = tempdir())
get_members_elected(2022)
options(op)
}
}
