% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate.R
\name{get_senate}
\alias{get_senate}
\title{Get Senate first preference votes by state}
\usage{
get_senate(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of Senate first preference votes by state.
}
\description{
Get Senate first preference votes by state
}
\examples{
\donttest{
op <- options(readaec.cache_dir = tempdir())
get_senate(2022)
options(op)
}
}
