% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{get_tcp}
\alias{get_tcp}
\title{Get two-candidate preferred votes by division}
\usage{
get_tcp(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of TCP votes by division.
}
\description{
Get two-candidate preferred votes by division
}
\examples{
\donttest{
get_tcp(2022)
}
}
