% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{get_tpp}
\alias{get_tpp}
\title{Get two-party preferred votes by division}
\usage{
get_tpp(year)
}
\arguments{
\item{year}{Election year. Use \code{\link[=list_elections]{list_elections()}} to see available years.}
}
\value{
A tidy data frame of TPP votes by division.
}
\description{
Get two-party preferred votes by division
}
\examples{
\donttest{
op <- options(readaec.cache_dir = tempdir())
get_tpp(2022)
options(op)
}
}
